Implemented ConnectStrava app
This consists of a simple app to handle the callback from Strava auth
page. All that is left to implement now is the post to SelfReflect
API/servers with the user id (state) and code. This requires the backend
to be completed so now currently waiting on that to be done. Should be
pretty simple to do once backend is done.