Implemented graph framework
All that is left to do now is implementation of the generate onClick
functionality. The functionality required is currently described in a
comment in the GenerateButton container. This must format the data is
fetches from the server in the correct way for the graph (LineChart) and
should handle all the collation of data, as well as
isLoading/isSubmitted updates, and updating data in state.